% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genpar.func}
\alias{genpar.func}
\title{Function to generate parameters given number of subjects n and rank parameter K.}
\usage{
genpar.func(n, K)
}
\arguments{
\item{n}{number of players/subjects.}

\item{K}{rank of parameter matrix.}
}
\value{
A list containing relevant parameters, including
\item{Pi}{Matrix of winning probabilities.}
\item{T_max}{Maximum number of comparisons between any pair.}
}
\description{
Function to generate parameters given number of subjects n and rank parameter K.
}
