% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{pcwst_est}
\alias{pcwst_est}
\title{Main estimation function}
\usage{
pcwst_est(Y, scale)
}
\arguments{
\item{Y}{a n x n matrix representing comparison outcomes.}

\item{scale}{Scale parameter corresponding to C_n in the paper.}
}
\value{
A list containing the result, including
\item{M_mat}{Matrix of estimated parameters.}
\item{iterations}{Number of iterations.}
\item{objective}{Value of objectvice function.}
\item{exit}{Exit status. 0 for successful convergence and 1 otherwise.}
}
\description{
Function to implement the proposed estimator based on given comparison matrix and scale parameter.
}
\examples{
# Generate example dataset
set.seed(123)
n=300
K=2
scale = 2*K
parameters <- genpar.func(n, K)
Y = gendata.func(T_max = parameters$T_max, Pi = parameters$Pi, n,  sparse_lv =1 )
result <-pcwst_est(Y,scale)
}
