% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{pcwst_est_tune}
\alias{pcwst_est_tune}
\title{Function to decide scale parameter controlling the nuclear norm constraint.}
\usage{
pcwst_est_tune(Y_train, Y_valid, scaleVals)
}
\arguments{
\item{Y_train}{A comparison matrix from the training set, used for estimating the model.}

\item{Y_valid}{A comparison matrix from the validation set, used for evaluating the likelihood.}

\item{scaleVals}{A numeric vector of candidate values for the tuning parameter.}
}
\value{
A list containing:
\item{out}{The model output corresponding to the optimal scale.}
\item{valid_like}{A numeric vector of validation likelihoods for each candidate scale.}
\item{scale}{The optimal scale value that maximizes the validation likelihood.}
}
\description{
Function to choose scale parameter based on training and validation data. Here the scale parameter is equivalent to Cn in the paper.
}
\examples{
set.seed(123)
n=300
K=2
scaleVals =10^seq(-1, 1, length.out = 20)
parameters = genpar.func(n, K)
Y_train = gendata.func(T_max = parameters$T_max, Pi = parameters$Pi, n,  sparse_lv =1 )
Y_valid = gendata.func(T_max = parameters$T_max, Pi = parameters$Pi, n,  sparse_lv =1 )
result = pcwst_est_tune(Y_train,Y_valid, scaleVals)

}
