% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{pcwst_gendata.func}
\alias{pcwst_gendata.func}
\title{Function to generate example data given T_max, winning probabilites, n and sparse level.}
\usage{
pcwst_gendata.func(T_max, Pi, n, sparse_lv, test = FALSE)
}
\arguments{
\item{T_max}{Maximum number of comparisons between any pair.}

\item{Pi}{Matrix of winning probabilities.}

\item{n}{number of players/subjects.}

\item{sparse_lv}{different levels of sparsity. Sparse(1), Less sparse (2) and dense (3).}
}
\value{
\item{Y}{Matrix of comparison outcome}
}
\description{
Function to generate example data given T_max, winning probabilites, n and sparse level.
}
